Loader created

+++ src/sat/Literal.java
@@ -6,7 +6,7 @@ package sat;
  * @author vitason
  */
 public class Literal extends Formula {
-    public Literal(boolean value, int weight) {
+    public Literal(boolean value, double weight) {
         super(value, weight);
     }
